Finally fixed the basic unit tests 🎉
Now that the basic unit tests have been prepared, I can try to get them working following the logic I added to the components.
My struggle was because Karma was kicking off some cryptic error about AuthService not being provided even though I was mocking it, the trick was about mocking also the AuthService in the `login.page.spec.ts` then providing it as a value.